fix(config): normalize required env values
