+++ .env.example
@@ -138,14 +138,14 @@ VIDEO_EXPORT_HOST_DIR=/home/capic/docker-data/dashboard-parapente/videos
 VIDEO_TEMP_IMAGES_HOST_DIR=/home/capic/docker-data/dashboard-parapente/video-temp-images
 
 # GoPro overlay toolchain and storage (required, no application default).
-# GOPRO_OVERLAY_HOST_DIR is the Docker host folder mounted at BACKEND_GOPRO_OVERLAY_ROOT.
-GOPRO_OVERLAY_HOST_DIR=/media/usb/data-m2/developement/gopro-overlay-dasboard
 BACKEND_GOPRO_OVERLAY_ROOT=/media/usb/data-m2/developement/gopro-overlay-dasboard
 BACKEND_GOPRO_OVERLAY_BIN=/media/usb/data-m2/developement/gopro-overlay-dasboard/venv/bin/gopro-dashboard.py
 BACKEND_GOPRO_OVERLAY_UPLOAD_DIR=/app/gopro-overlays/uploads
 BACKEND_GOPRO_OVERLAY_OUTPUT_DIR=/app/gopro-overlays/outputs
 BACKEND_GOPRO_OVERLAY_PARAGLIDING_ROOT=/app/gopro-overlays/paragliding
 BACKEND_GOPRO_OVERLAY_LAYOUT_DIR=/media/usb/data-m2/developement/gopro-overlay-dasboard
+# GOPRO_OVERLAY_HOST_DIR is the Docker host folder mounted at BACKEND_GOPRO_OVERLAY_ROOT.
+GOPRO_OVERLAY_HOST_DIR=/media/usb/data-m2/developement/gopro-overlay-dasboard
 
 # Deployment version state file (required, no application default).
 BACKEND_VERSION_STATE_FILE=/app/db/version_state.json
