+++ apps/backend/config.py
@@ -13,15 +13,9 @@ from pathlib import Path
 
 from dotenv import load_dotenv
 
+from env_utils import required_env
+
 logger = logging.getLogger(__name__)
-
-
-def required_env(name: str) -> str:
-    value = os.getenv(name)
-    if value is None or not value.strip():
-        logger.error("Required environment variable %s is missing or empty", name)
-        raise ValueError(f"{name} environment variable is required")
-    return value
 
 
 # Déterminer le répertoire racine du projet
